Title: 사각형 채우기 3, Time: 6ms, Memory: 0MB, Status: Passed - Codetree

+++ 250428/사각형 채우기 3/rectangle-fill-3.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int n;
-int dp[1001];
+long long dp[1001];
 
 int main() {
     cin >> n;
